+++ src/main/java/com/jj/Gradebook/dao/TimetableRepository.java
@@ -3,5 +3,5 @@ package com.jj.Gradebook.dao;
 import com.jj.Gradebook.entity.Timetable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TimetableRepository extends JpaRepository<Timetable, Integer> {
+public interface TimetableRepository extends JpaRepository<Timetable, Long> {
 }
